Rename utils hpp header
